[filament_scene] Material upgrades, camera widget, light fixes

+++ packages/filament_scene/example/lib/demo_widgets.dart
@@ -122,7 +122,7 @@ class _LightSettingsWidgetState extends State<LightSettingsWidget> {
                     widget.filament.changeLightColorByGUID(
                       centerPointLightGUID,
                       colorString,
-                      _directIntensity.toInt(),
+                      _directIntensity,
                     ),
                   );
                 });
@@ -156,7 +156,7 @@ class _LightSettingsWidgetState extends State<LightSettingsWidget> {
                         widget.filament.changeLightColorByGUID(
                           centerPointLightGUID,
                           colorString,
-                          _directIntensity.toInt(),
+                          _directIntensity,
                         ),
                       );
                     });
